awsserver/inference/requirements.txt
pandas==1.1.4
numpy==1.19.3
sentence_transformers==2.0.0
faiss-cpu